pyproject.toml
[project]
name = "charter"
version = "0.1.0"
description = "Charter offers cache service"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.135,<0.136",
    "uvicorn>=0.42,<0.43",
    "sqlalchemy>=2.0,<2.1",
    "asyncpg>=0.31,<0.32",
    "alembic>=1.18,<1.19",
    "pydantic-settings>=2.13,<3.0",
    "pyjwt[crypto]>=2.8,<3.0",
    "bcrypt>=4.0,<5.0",
    "email-validator>=2.3.0",
]

[project.scripts]
dev = "app.main:run_dev"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]

ignore = [
    "C408",
    "RUF001", "RUF002", "RUF003",
    "B008",   # FastAPI: Depends() в аргументах — это норма для FastAPI
    "S105",   # token_type = "bearer" — это не пароль
    "E501",   # длинные строки — решайте вручную по необходимости
]

fixable = [
    "I",
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"migration/*" = ["UP007", "UP035", "I001", "ERA001", "E501", "F401", "W291"]
